+++ CHANGELOG.md
@@ -21,6 +21,11 @@
 - iOS, tvOS: `onFairplayLicenseAcquired` player event, providing the `contentKeyRequest` of the acquired FairPlay license
 - iOS, tvOS: `Source.drm.fairplay.renewExpiringLicense` method to proactively renew an expiring FairPlay license using the `contentKeyRequest` from a `FairplayLicenseAcquiredEvent`
 
+### Known Issues
+
+- iOS: Moving a live `PlayerView` between different React Native view trees while using IMA ads in the same source is unsupported and can result in a black screen during playback
+  - A native warning log is now emitted when a rebind of this type is detected
+
 ## [1.9.0] - 2026-02-13
 
 ### Changed